% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqQuantile.R
\name{Arnold}
\alias{Arnold}
\title{Functional form of Lorenz curve : Arnold (1986)}
\usage{
Arnold(p, a, b)
}
\arguments{
\item{p}{numeric, vector of probabilities}

\item{a}{numeric, positive}

\item{b}{numeric, positive}
}
\value{
The Arnold (1986)'s functional form of Lorenz curve at the point \code{p} \deqn{L(p) =  (p*(1 + (a-1)*p))/(1 + (a-1)*p + b*(1-p))}  where \eqn{a, b > 0} and \eqn{a - b < 1}.
}
\description{
Functional form of Lorenz curve : Arnold (1986)
}
\examples{
Arnold(0.5, 1.2, 2.2)
0.3333333
}
\references{
Arnold (1986), \emph{A class of hyperbolic Lorenz curves}
}
\seealso{
\code{\link{KP}}, \code{\link{RGKO}}, \code{\link{Ortega}}, \code{\link{Chotikapanich}}, \code{\link{Sarabia}}, \code{\link{Rohde}}
}
