% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqQuantile.R
\name{compute_LC}
\alias{compute_LC}
\title{Means of the bins according the conditional expectation or midpoint methods}
\usage{
compute_LC(ID, p, bound_min, bound_max, nb, method = "CondExp",
  whichpareto = 0.8)
}
\arguments{
\item{ID}{vector of ID area}

\item{p}{numeric, vector of probabilities}

\item{bound_min}{numeric, minimum of the interval}

\item{bound_max}{numeric, maximum of the interval}

\item{nb}{numeric, number of the interval}

\item{method}{string, type of methods ("CondExp" for conditional expectation method of "Midpoint" for midpoint method)}

\item{whichpareto}{numeric, probability from which pareto tail is assumed}
}
\value{
A dataframe with the bounds, the mean, the cumulative income share and the cumulative population share.
}
\description{
Means of the bins according the conditional expectation or midpoint methods
}
\examples{
data("tabulated_income")
BelAir5 = tabulated_income[tabulated_income$iris=="Bel Air 5",]
compute_LC(ID=BelAir5$iris, p=BelAir5$prop_cum_population, bound_min = BelAir5$bound_min, bound_max = BelAir5$bound_max, nb = BelAir5$prop_population, method = "CondExp")
compute_LC(ID=BelAir5$iris, p=BelAir5$prop_cum_population, bound_min = BelAir5$bound_min, bound_max = BelAir5$bound_max, nb = BelAir5$prop_population, method = "Midpoint")
}
\references{
Belz (2019), \emph{Estimating Inequality Measure from Quantile Data}
}
\seealso{
\code{\link{run_compute_LC}}
}
