% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqQuantile.R
\name{compute_Pietra}
\alias{compute_Pietra}
\title{Computation of Pietra index according to a functional form and its parameters}
\usage{
compute_Pietra(function_form, par1, par2, par3 = NA, par4 = NA)
}
\arguments{
\item{function_form}{string, functional form in "KP", "RGKO", "ARNOLD", "CHOTIKAPANICH", "SARABIA", "ORTEGA" and "ROHDE"}

\item{par1}{numeric, parameter of the functional form}

\item{par2}{numeric, parameter of the functional form}

\item{par3}{numeric, parameter of the functional form}

\item{par4}{numeric, parameter of the functional form}
}
\value{
The value of the Pietra index accorting a functional form and its parameters
}
\description{
Computation of Pietra index according to a functional form and its parameters
}
\examples{
data("tabulated_income")
LC_tabulated_income = run_compute_LC(ID=tabulated_income$ID, p=tabulated_income$prop_cum_population, bound_min = tabulated_income$bound_min, bound_max = tabulated_income$bound_max, nb = tabulated_income$prop_population, method = "CondExp")
Optim_LC = run_optim_LC(ID = unique(LC_tabulated_income$ID),income_cum = LC_tabulated_income$income_cum, population_cum=LC_tabulated_income$population_cum, function_form = c("KP", "RGKO", "ARNOLD", "CHOTIKAPANICH", "SARABIA", "ORTEGA", "ROHDE"))
Optim_LC
compute_Pietra(function_form = "KP", par1 = 1, par2 = 1.450156)
0.2645622
Optim_LC \%>\% ungroup() \%>\%  rowwise() \%>\% mutate(Pietra = compute_Pietra(function_form, par1, par2, par3, par4)) \%>\% select(ID, function_form, Pietra)
}
\seealso{
\code{\link{compute_Gini}}, \code{\link{compute_TH}}, \code{\link{compute_TL}}, \code{\link{compute_topshare}}
}
