% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqQuantile.R
\name{Gini_Arnold}
\alias{Gini_Arnold}
\title{Gini of the functional form  Arnold (1986)}
\usage{
Gini_Arnold(a, b)
}
\arguments{
\item{a}{numeric}

\item{b}{numeric}
}
\value{
The value of the Gini index
}
\description{
Gini of the functional form  Arnold (1986)
}
\examples{
Gini_Arnold(1.2, 2.2)
0.3484886
}
\references{
Arnold (1986), \emph{A class of hyperbolic Lorenz curves}
}
\seealso{
\code{\link{Gini_KP}}, \code{\link{Gini_RGKO}}, \code{\link{Gini_Ortega}},  \code{\link{Gini_Chotikapanich}}, \code{\link{Gini_Sarabia}}, \code{\link{Gini_Rohde}}
}
