% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqQuantile.R
\name{Gini_Chotikapanich}
\alias{Gini_Chotikapanich}
\title{Gini of the functional form  Chotikapanich (1993)}
\usage{
Gini_Chotikapanich(k)
}
\arguments{
\item{k}{numeric}
}
\value{
The value of the Gini index
}
\description{
Gini of the functional form  Chotikapanich (1993)
}
\examples{
Gini_Chotikapanich(2.2)
0.3401299
}
\references{
Chotikapanich (1993),  \emph{A comparison of alternative functional forms for the Lorenz curve}
}
\seealso{
\code{\link{Gini_KP}}, \code{\link{Gini_RGKO}}, \code{\link{Gini_Arnold}},  \code{\link{Gini_Ortega}}, \code{\link{Gini_Sarabia}}, \code{\link{Gini_Rohde}}
}
