% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqQuantile.R
\name{Gini_KP}
\alias{Gini_KP}
\title{Gini of the functional form  Kakwani and Podder (1973)}
\usage{
Gini_KP(a, b)
}
\arguments{
\item{a}{numeric}

\item{b}{numeric}
}
\value{
The value of the Gini index
}
\description{
Gini of the functional form  Kakwani and Podder (1973)
}
\examples{
Gini_KP(0.9, 1.5)
0.3331993
}
\references{
Kakwani and Podder (1973), \emph{On the estimation of Lorenz curves from grouped observations}
}
\seealso{
\code{\link{Gini_RGKO}}, \code{\link{Gini_Arnold}}, \code{\link{Gini_Ortega}},  \code{\link{Gini_Chotikapanich}}, \code{\link{Gini_Sarabia}}, \code{\link{Gini_Rohde}}
}
