% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqQuantile.R
\name{Gini_Rohde}
\alias{Gini_Rohde}
\title{Gini of the functional form  Rohde (2009)}
\usage{
Gini_Rohde(b)
}
\arguments{
\item{b}{numeric, greater than 1}
}
\value{
The value of the Gini index of the functional form : Rohde (2009)
}
\description{
Gini of the functional form  Rohde (2009)
}
\examples{
Gini_Rohde(1.5)
0.3520816
}
\references{
Rohde (2009), \emph{An alternative functional form for estimating the Lorenz curve}
}
\seealso{
\code{\link{Gini_KP}}, \code{\link{Gini_RGKO}}, \code{\link{Gini_Arnold}},  \code{\link{Gini_Ortega}}, \code{\link{Gini_Chotikapanich}}, \code{\link{Gini_Sarabia}}
}
