% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqQuantile.R
\name{Gini_Sarabia}
\alias{Gini_Sarabia}
\title{Gini of the functional form  Sarabia (1997)}
\usage{
Gini_Sarabia(pi1, pi2, a1, a2)
}
\arguments{
\item{pi1}{numeric}

\item{pi2}{numeric}

\item{a1}{numeric}

\item{a2}{numeric}
}
\value{
The value of the Gini index
}
\description{
Gini of the functional form  Sarabia (1997)
}
\examples{
Gini_Sarabia(0.1, 0.7, 1.8, 0.3)
0.3076923
}
\references{
Sarabia (1997), \emph{A hierarchy of Lorenz curves based on the generalized Tukey’s lambda distribution}
}
\seealso{
\code{\link{Gini_KP}}, \code{\link{Gini_RGKO}}, \code{\link{Gini_Arnold}},  \code{\link{Gini_Ortega}}, \code{\link{Gini_Chotikapanich}}, \code{\link{Gini_Rohde}}
}
