% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqQuantile.R
\name{KP}
\alias{KP}
\title{Functional form Kakwani and Podder (1973)}
\usage{
KP(p, a, b)
}
\arguments{
\item{p}{numeric, vector of probabilities}

\item{a}{numeric}

\item{b}{numeric}
}
\value{
The value of the Lorenz curve at the point \code{p}
}
\description{
Functional form Kakwani and Podder (1973)
}
\examples{
KP(0.5, 0.9, 1.5)
0.253135
}
\references{
Kakwani and Podder (1973), \emph{On the estimation of Lorenz curves from grouped observations}
}
\seealso{
\code{\link{RGKO}}, \code{\link{Arnold}}, \code{\link{Ortega}}, \code{\link{Chotikapanich}}, \code{\link{Sarabia}}, \code{\link{Rohde}}
}
