% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqQuantile.R
\name{Lprime}
\alias{Lprime}
\title{Computation of the derivative of a function L at the point x}
\usage{
Lprime(x, L, h = 1e-05)
}
\arguments{
\item{x}{numeric, point}

\item{L}{function}

\item{h}{numeric}
}
\description{
Computation of the derivative of a function L at the point x
}
