% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqQuantile.R
\name{optim_LC}
\alias{optim_LC}
\title{Optimisation of a parametric Lorenz curve}
\usage{
optim_LC(ID, income_cum, population_cum, function_form)
}
\arguments{
\item{ID}{vector of ID area}

\item{income_cum}{numeric, vector of cumulaive income shares}

\item{population_cum}{numeric, vector of cumulative population shares}

\item{function_form}{string, functional form in "KP", "RGKO", "ARNOLD", "CHOTIKAPANICH", "SARABIA", "ORTEGA" and "ROHDE"}
}
\value{
A dataframe with the functional forms, the parameters, the value of the NLS and the Chi-squared statistic.
}
\description{
Optimisation of a parametric Lorenz curve
}
\references{
Belz (2019), \emph{Estimating Inequality Measures from Quantile Data} \url{https://halshs.archives-ouvertes.fr/halshs-02320110}
}
\seealso{
\code{\link{run_optim_LC}}
}
