% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqQuantile.R
\name{optim_LC_function}
\alias{optim_LC_function}
\title{Optimisation of a parametric Lorenz curve for several functional forms}
\usage{
optim_LC_function(ID, income_cum, population_cum, function_form)
}
\arguments{
\item{ID}{vector of ID area}

\item{income_cum}{numeric, vector of cumulaive income shares}

\item{population_cum}{numeric, vector of cumulative population shares}

\item{function_form}{string, functional form in "KP", "RGKO", "ARNOLD", "CHOTIKAPANICH", "SARABIA", "ORTEGA" and "ROHDE"}
}
\value{
A dataframe with functional form, the parameters, the value of the NLS and the Chi-squared statistic
}
\description{
Optimisation of a parametric Lorenz curve for several functional forms
}
\seealso{
\code{\link{run_optim_LC}}
}
