% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqQuantile.R
\name{Ortega}
\alias{Ortega}
\title{Functional form of Lorenz curve : Ortega et al. (1991)}
\usage{
Ortega(p, a, b)
}
\arguments{
\item{p}{numeric, vector of probabilities}

\item{a}{numeric, positive}

\item{b}{numeric, between 0 and 1}
}
\value{
The Ortega et al. (1991)'s functional form of Lorenz curve at the point \code{p} \deqn{L(p) =  p^a*(1 - (1-p)^b)}  where \eqn{a \ge 0} and \eqn{0 < b <1 }.
}
\description{
Functional form of Lorenz curve : Ortega et al. (1991)
}
\examples{
Ortega(0.5, 0.5, 0.7)
0.2718315
}
\references{
Ortega et al. (1991), \emph{A new functional form for estimating Lorenz curves}
}
\seealso{
\code{\link{KP}}, \code{\link{RGKO}}, \code{\link{Arnold}},  \code{\link{Chotikapanich}}, \code{\link{Sarabia}}, \code{\link{Rohde}}
}
