% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqQuantile.R
\name{run_compute_LC}
\alias{run_compute_LC}
\title{Means of the bins according the conditional expectation or midpoint methods for several areas}
\usage{
run_compute_LC(ID, p, bound_min, bound_max, nb, method = "CondExp",
  whichpareto = 0.8)
}
\arguments{
\item{ID}{vector of ID area}

\item{p}{numeric, vector of probabilities}

\item{bound_min}{numeric, minimum of the interval}

\item{bound_max}{numeric, maximum of the interval}

\item{nb}{numeric, number of the interval}

\item{method}{string, type of methods ("CondExp" for conditional expectation method of "Midpoint" for midpoint method)}

\item{whichpareto}{numeric, probability from which pareto tail is assumed}
}
\value{
A dataframe with the bounds, the mean, the cumulative income share and the cumulative population share.
}
\description{
Means of the bins according the conditional expectation or midpoint methods for several areas
}
\examples{
data("tabulated_income")
run_compute_LC(ID=tabulated_income$iris, p=tabulated_income$prop_cum_population, bound_min = tabulated_income$bound_min, bound_max = tabulated_income$bound_max, nb = tabulated_income$prop_population, method = "CondExp")
}
\references{
Belz (2019), \emph{Estimating Inequality Measure from Quantile Data}
}
