% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqQuantile.R
\name{run_optim_LC}
\alias{run_optim_LC}
\title{Optimisation of a parametric Lorenz curve for several functional forms and several areas}
\usage{
run_optim_LC(ID, income_cum, population_cum, function_form)
}
\arguments{
\item{ID}{vector of ID area}

\item{income_cum}{numeric, vector of cumulaive income shares}

\item{population_cum}{numeric, vector of cumulative population shares}

\item{function_form}{string, functional form in "KP", "RGKO", "ARNOLD", "CHOTIKAPANICH", "SARABIA", "ORTEGA" and "ROHDE"}
}
\value{
A dataframe with functional form, the parameters, the value of the NLS and the Chi-squared statistic
}
\description{
Optimisation of a parametric Lorenz curve for several functional forms and several areas
}
\examples{
data("tabulated_income")
LC_tabulated_income = run_compute_LC(ID=tabulated_income$ID, p=tabulated_income$prop_cum_population, bound_min = tabulated_income$bound_min, bound_max = tabulated_income$bound_max, nb = tabulated_income$prop_population, method = "CondExp")
run_optim_LC(ID = unique(LC_tabulated_income$ID),income_cum = LC_tabulated_income$income_cum, population_cum=LC_tabulated_income$population_cum, function_form = c("KP", "RGKO", "ARNOLD", "CHOTIKAPANICH", "SARABIA", "ORTEGA", "ROHDE"))
}
