% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqQuantile.R
\name{Sarabia}
\alias{Sarabia}
\title{Functional form of Lorenz curve :  Sarabia (1997)}
\usage{
Sarabia(p, pi1, pi2, a1, a2)
}
\arguments{
\item{p}{numeric, vector of probabilities}

\item{pi1}{numeric, between 0 and 1}

\item{pi2}{numeric, between 0 and 1}

\item{a1}{numeric, greater than 1}

\item{a2}{numeric, between 0 and 1}
}
\value{
The  Sarabia (1997)'s functional form of Lorenz curve at the point \code{p} \deqn{L(p) =  pi1*p + pi2*p^a1 + (1 - pi1 - pi2)*(1 - (1-p)^a2)}  where \eqn{0 \le pi1, pi2 \le 1, a1 \ge  1} and \eqn{0 < a2 < 1}.
}
\description{
Functional form of Lorenz curve :  Sarabia (1997)
}
\examples{
Sarabia(0.5, 0.1, 0.7, 1.8, 0.3)
0.2885717
}
\references{
Sarabia (1997), \emph{A hierarchy of Lorenz curves based on the generalized Tukey’s lambda distribution}
}
\seealso{
\code{\link{KP}}, \code{\link{RGKO}}, \code{\link{Arnold}},  \code{\link{Ortega}}, \code{\link{Chotikapanich}}, \code{\link{Rohde}}
}
