% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulated_income.R
\docType{data}
\name{tabulated_income}
\alias{tabulated_income}
\title{Tabulated data of Bel Air 5 and Combat 7 iris for the year 2014}
\format{A data frame with 24 rows and 5 variables:
\describe{
  \item{iris}{Name of the area}
  \item{bound_min}{Minimum of the income range}
  \item{bound_max}{Maximum of the income range}
  \item{prop_population}{Population share}
  \item{prop_cum_population}{Cumulative population share}
  ...
}}
\source{
\url{https://www.insee.fr/fr/statistiques/3288151}
}
\usage{
tabulated_income
}
\description{
Tabulated data of Bel Air 5 and Combat 7 iris for the year 2014
}
\keyword{datasets}
