% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqQuantile.R
\name{Arnold}
\alias{Arnold}
\title{Functional form  Arnold (1986)}
\usage{
Arnold(p, a, b)
}
\arguments{
\item{p}{numeric, vector of probabilities}

\item{a}{numeric}

\item{b}{numeric}
}
\value{
The value of the Lorenz curve at the point \code{p}
}
\description{
Functional form  Arnold (1986)
}
\examples{
Arnold(0.5, 1.2, 2.2)
0.3333333
}
\references{
Arnold (1986), \emph{A class of hyperbolic Lorenz curves}
}
\seealso{
\code{\link{KP}}, \code{\link{RGKO}}, \code{\link{Ortega}}, \code{\link{Chotikapanich}}, \code{\link{Sarabia}}, \code{\link{Rohde}}
}
