% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqQuantile.R
\name{Chotikapanich}
\alias{Chotikapanich}
\title{Functional form  Chotikapanich (1993)}
\usage{
Chotikapanich(p, k)
}
\arguments{
\item{p}{numeric, vector of probabilities}

\item{k}{numeric}
}
\value{
The value of the Lorenz curve at the point \code{p}
}
\description{
Functional form  Chotikapanich (1993)
}
\examples{
Chotikapanich(0.5, 2.2)
0.2497399
}
\references{
Chotikapanich (1993), \emph{A comparison of alternative functional forms for the Lorenz curve}
}
\seealso{
\code{\link{KP}}, \code{\link{RGKO}}, \code{\link{Arnold}},  \code{\link{Ortega}}, \code{\link{Sarabia}}, \code{\link{Rohde}}
}
