% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqQuantile.R
\name{Gini_Ortega}
\alias{Gini_Ortega}
\title{Gini of the functional form  Ortega et al. (1991)}
\usage{
Gini_Ortega(a, b)
}
\arguments{
\item{a}{numeric}

\item{b}{numeric}
}
\value{
The value of the Gini index
}
\description{
Gini of the functional form  Ortega et al. (1991)
}
\examples{
Gini_Ortega(0.5, 0.7)
0.3310822
}
\references{
Ortega et al. (1991),  \emph{A new functional form for estimating Lorenz curves}
}
\seealso{
\code{\link{Gini_KP}}, \code{\link{Gini_RGKO}}, \code{\link{Gini_Arnold}},  \code{\link{Gini_Chotikapanich}}, \code{\link{Gini_Sarabia}}, \code{\link{Gini_Rohde}}
}
