% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqQuantile.R
\name{Gini_RGKO}
\alias{Gini_RGKO}
\title{Gini of the functional form  Rasche et al. (1980)}
\usage{
Gini_RGKO(a, b)
}
\arguments{
\item{a}{numeric}

\item{b}{numeric}
}
\value{
The value of the Gini index
}
\description{
Gini of the functional form  Rasche et al. (1980)
}
\examples{
Gini_RGKO(0.7, 1.5)
0.3868911
}
\references{
Rasche et al. (1980), \emph{Functional forms for estimating the Lorenz curve: comment}
}
\seealso{
\code{\link{Gini_KP}}, \code{\link{Gini_Arnold}}, \code{\link{Gini_Ortega}},  \code{\link{Gini_Chotikapanich}}, \code{\link{Gini_Sarabia}}, \code{\link{Gini_Rohde}}
}
