% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqQuantile.R
\name{Ortega}
\alias{Ortega}
\title{Functional form  Ortega et al. (1991)}
\usage{
Ortega(p, a, b)
}
\arguments{
\item{p}{numeric, vector of probabilities}

\item{a}{numeric}

\item{b}{numeric}
}
\value{
The value of the Lorenz curve at the point \code{p}
}
\description{
Functional form  Ortega et al. (1991)
}
\examples{
Ortega(0.5, 0.5, 0.7)
0.2718315
}
\references{
Ortega et al. (1991), \emph{A new functional form for estimating Lorenz curves}
}
\seealso{
\code{\link{KP}}, \code{\link{RGKO}}, \code{\link{Arnold}},  \code{\link{Chotikapanich}}, \code{\link{Sarabia}}, \code{\link{Rohde}}
}
