% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqQuantile.R
\name{Rohde}
\alias{Rohde}
\title{Functional form  Rohde (2009)}
\usage{
Rohde(p, b)
}
\arguments{
\item{p}{numeric, vector of probabilities}

\item{b}{numeric}
}
\value{
The value of the Lorenz curve at the point \code{p}
}
\description{
Functional form  Rohde (2009)
}
\examples{
Rohde(0.5, 1.5)
0.25
}
\references{
Rohde (2009), \emph{An alternative functional form for estimating the Lorenz curve}
}
\seealso{
\code{\link{KP}}, \code{\link{RGKO}}, \code{\link{Arnold}},  \code{\link{Ortega}}, \code{\link{Chotikapanich}}, \code{\link{Sarabia}}
}
