% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqQuantile.R
\name{Sarabia}
\alias{Sarabia}
\title{Functional form  Sarabia (1997)}
\usage{
Sarabia(p, pi1, pi2, a1, a2)
}
\arguments{
\item{p}{numeric, vector of probabilities}

\item{pi1}{numeric}

\item{pi2}{numeric}

\item{a1}{numeric}

\item{a2}{numeric}
}
\value{
The value of the Lorenz curve at the point \code{p}
}
\description{
Functional form  Sarabia (1997)
}
\examples{
Sarabia(0.5, 0.1, 0.7, 1.8, 0.3)
0.2885717
}
\references{
Sarabia (1997), \emph{A hierarchy of Lorenz curves based on the generalized Tukey’s lambda distribution}
}
\seealso{
\code{\link{KP}}, \code{\link{RGKO}}, \code{\link{Arnold}},  \code{\link{Ortega}}, \code{\link{Chotikapanich}}, \code{\link{Rohde}}
}
